requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiohttp==3.9.0
google-generativeai==0.3.0
openai==0.28.0
python-json-logger==2.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
python-multipart==0.0.6
